Pass expected size into diff functions
We were previously assuming that the output buffer should be the same
size as the data input buffer. This was causing overflows as most
patches contain compression logic, so we'd expect the output buffer to
be bigger when decompressed.

We now calculate the expected final size and pass this so that an output
buffer can be allocated of this size.