Remove copied code in non-delta functions
A bunch of the cases: ZERO, REPLACE, REPLACE_XZ, and REPLACE_BZ all do
the same logic - copy the desired data as-is over to the destination
blocks. The only difference is how the data is encoded.

We have now combined these and instead we simply swap out the io.Reader
based on the input encoding.