Exit early on hash mismatches
Rather than pushing through, as we previously did, we now specifically
stop when hashes do not match. There seems very little benefit in
applying a patch that we are fairly certain at this point will not work,
and it is possible this may cause memory bugs if our output buffer is
initialised to the wrong length.

This commit also improves the error message slightly. When running split
across threads, this error can show multiple times, but this can be
fixed at a later time.